+++ app/components/ui/index.ts
@@ -6,7 +6,6 @@ export * from './datepicker'
 export * from './dropdown'
 export * from './expandable-arrow'
 export * from './image-tile'
-export * from './paginated-content'
 export * from './pagination'
 export * from './select'
 export * from './shallow-tile'
