+++ app/features/dynamic-content/components/dynamic-content-empty-state.tsx
@@ -1,0 +1,6 @@
+import React, {FC, PropsWithChildren} from 'react'
+import {Text} from '@effortless-ui'
+
+export const DynamicContentEmptyState: FC<PropsWithChildren> = ({children}) => {
+  return <Text>{children}</Text>
+}
