+++ app/features/dynamic-content/types.ts
@@ -1,0 +1,8 @@
+import {IApiFilters, TApiCommonItem} from '@/app/features/api/types'
+
+export * from './components/dynamic-content'
+
+export interface ISelectedFilter<T extends TApiCommonItem> {
+  filters: IApiFilters<T>[]
+  label: string
+}
