+++ pages/news/index.tsx
@@ -2,9 +2,10 @@ import React from 'react'
 import {GetServerSideProps, NextPage} from 'next'
 
 import MasterPage from '@/app/components/masterpages/masterpage'
-import {PaginatedContent, Tile, TilesList, Title} from '@/app/components/ui'
+import {Tile, TilesList, Title} from '@/app/components/ui'
 import {TApiNews} from '@/app/features/api/types'
 import {getDehydratedState, IGetApiCollectionResponseParams, IPageWithPayload} from '@/app/features/api/utils'
+import {DynamicContent} from '@/app/features/dynamic-content'
 import {useScrollRef} from '@/app/hooks'
 import {theme} from '@/app/styles'
 import {mapApiNewsToTile} from '@/app/utils'
@@ -17,7 +18,7 @@ const Page: NextPage<IPageWithPayload<[TApiNews]>> = ({payloads: [payload]}) => 
       <Title ref={scrollRef} cs={{mb: [theme.spacing.l, theme.spacing.xxl]}}>
         Aktualności
       </Title>
-      <PaginatedContent payload={payload} scrollToElement={scrollToElement}>
+      <DynamicContent payload={payload} scrollToElement={scrollToElement}>
         {(data) => (
           <TilesList
             cols={[1, 2, 3]}
@@ -26,7 +27,7 @@ const Page: NextPage<IPageWithPayload<[TApiNews]>> = ({payloads: [payload]}) => 
             ))}
           />
         )}
-      </PaginatedContent>
+      </DynamicContent>
     </MasterPage>
   )
 }
